--- templates/apps/item-usage-flow.hbs
@@ -1,70 +1,0 @@
-<p>{{title}}</p>
-
-<form id="ability-use-form">
-  {{#if methods.canConsumeSlot}}
-  <div class="form-group">
-    <label>{{ localize "DND5E.SpellCastUpcast" }}</label>
-    <div class="form-fields">
-      <select name="currentSlot">
-        {{#select config.currentSlot}}
-        {{#each spellLevels as |l|}}
-        <option value="{{l.level}}" {{disabled (not l.canCast)}}>{{l.label}}</option>
-        {{/each}}
-        {{/select}}
-      </select>
-    </div>
-  </div>
-
-  <div class="form-group">
-    <label class="checkbox">
-      <input type="checkbox" name="consumeSlot" {{checked config.consumeSlot}}>{{localize "DND5E.SpellCastConsume"}}
-    </label>
-  </div>
-  {{/if}}
-
-  {{#if methods.canConsumeUses}}
-  <div class="form-group">
-    <label class="checkbox">
-      <input type="checkbox" name="consumeUsage" {{checked config.consumeUsage}}>{{localize "DND5E.AbilityUseConsume"}}
-    </label>
-  </div>
-  {{/if}}
-
-  {{#if methods.canConsumeResource}}
-  <div class="form-group">
-    <label class="checkbox">
-      <input type="checkbox" name="consumeResource" {{checked config.consumeResource}}>{{localize "DND5E.ConsumeResource"}}
-    </label>
-  </div>
-  {{/if}}
-
-  {{#if methods.canCreateTemplate}}
-  <div class="form-group">
-    <label class="checkbox">
-      <input type="checkbox" name="createMeasuredTemplate" {{checked config.createMeasuredTemplate}}>{{localize "DND5E.PlaceTemplate"}}
-    </label>
-  </div>
-  {{/if}}
-
-  <footer>
-    <button data-type="submit" {{disabled disabled}}>
-      <i class="fa-solid {{icon}}"></i> {{localize label}}
-    </button>
-  </footer>
-</form>
-
-<hr>
-
-<div class="notes">
-  {{#each notes}}
-  <p class="note">{{this}}</p>
-  {{/each}}
-</div>
-
-<hr>
-
-<div class="warnings">
-  {{#each warnings}}
-  <p class="notification warning">{{this}}</p>
-  {{/each}}
-</div>
